refactor(ffi): move the `TimelineEventTypeFilter` to `timeline/configuration`
Pure code motion, nothing else.
